// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/bleeding-entry-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface BleedingEntryHttpResponse {
    cycleProfileId?: string;
    date?: string;
    flow?: number;
    id?: string;
    notes?: string | null;
    painImpact?: number | null;
    type?: number;
}