// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/cycle-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { BleedingEntryHttpResponse } from './bleeding-entry-http-response';
import { CycleConsentHttpResponse } from './cycle-consent-http-response';
import { CycleDayNoteHttpResponse } from './cycle-day-note-http-response';
import { CycleFactorHttpResponse } from './cycle-factor-http-response';
import { CyclePredictionRevisionHttpResponse } from './cycle-prediction-revision-http-response';
import { CyclePredictionsHttpResponse } from './cycle-predictions-http-response';
import { CycleSymptomEntryHttpResponse } from './cycle-symptom-entry-http-response';
import { FertilitySignalHttpResponse } from './fertility-signal-http-response';
import { MenstrualEpisodeHttpResponse } from './menstrual-episode-http-response';

export interface CycleHttpResponse {
    averageCycleLength?: number;
    averagePeriodLength?: number;
    bleedingEntries?: Array<BleedingEntryHttpResponse>;
    confidence?: number;
    consents?: Array<CycleConsentHttpResponse> | null;
    dayNotes?: Array<CycleDayNoteHttpResponse> | null;
    discreetNotifications?: boolean;
    factors?: Array<CycleFactorHttpResponse>;
    fertilitySignals?: Array<FertilitySignalHttpResponse>;
    goal?: number;
    hideFromDashboard?: boolean;
    id?: string;
    isOnboardingComplete?: boolean;
    isRegular?: boolean;
    lutealLength?: number;
    menstrualEpisodes?: Array<MenstrualEpisodeHttpResponse>;
    mode?: number;
    notes?: string | null;
    predictionRevisions?: Array<CyclePredictionRevisionHttpResponse> | null;
    predictions?: CyclePredictionsHttpResponse | null;
    reproductiveState?: number;
    showFertilityEstimates?: boolean;
    symptoms?: Array<CycleSymptomEntryHttpResponse>;
    trackingStartDate?: string;
    userId?: string;
}