// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/cycle-log-day-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { BleedingEntryHttpResponse } from './bleeding-entry-http-response';
import { CycleSymptomEntryHttpResponse } from './cycle-symptom-entry-http-response';
import { FertilitySignalHttpResponse } from './fertility-signal-http-response';

export interface CycleLogDayHttpResponse {
    bleedingEntries?: Array<BleedingEntryHttpResponse>;
    cycleProfileId?: string;
    date?: string;
    fertilitySignal?: FertilitySignalHttpResponse | null;
    notes?: string | null;
    symptoms?: Array<CycleSymptomEntryHttpResponse>;
}