// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/cycle-nutrition-summary-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface CycleNutritionSummaryHttpResponse {
    algorithmVersion?: string;
    averageCaloriesOnBleedingDays?: number;
    averageCaloriesOnNonBleedingCycleDays?: number;
    averageFiberOnBleedingDays?: number;
    averageFiberOnNonBleedingCycleDays?: number;
    averagePainImpactOnDaysWithMeals?: number;
    bleedingDays?: number;
    comparableCycles?: number;
    completedCyclesAnalyzed?: number;
    consentRequired?: boolean;
    dataSufficiency?: string;
    dateFrom?: string;
    dateTo?: string;
    daysWithMeals?: number;
    hasEnoughNutritionData?: boolean;
    loggedCycleDays?: number;
    reasonCodes?: Array<string> | null;
}