// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/cycle-prediction-revision-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface CyclePredictionRevisionHttpResponse {
    algorithmVersion?: string;
    calibrationSampleCount?: number;
    completedCycleCount?: number;
    confidence?: string;
    dataSufficiency?: string;
    generatedAtUtc?: string;
    historicalCoveragePercent?: number | null;
    id?: string;
    meanAbsoluteErrorDays?: number | null;
    nextPeriodStartFrom?: string | null;
    nextPeriodStartTo?: string | null;
    patternConsistency?: string;
    reasonCodes?: Array<string>;
}