// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/cycle-symptom-entry-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface CycleSymptomEntryHttpResponse {
    category?: number;
    cycleProfileId?: string;
    date?: string;
    id?: string;
    intensity?: number;
    note?: string | null;
    tags?: Array<string>;
}