// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/fertility-signal-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface FertilitySignalHttpResponse {
    basalBodyTemperatureCelsius?: number | null;
    cervicalFluid?: string | null;
    cycleProfileId?: string;
    date?: string;
    hadSex?: boolean | null;
    id?: string;
    notes?: string | null;
    ovulationTestResult?: number | null;
}